Fix GetAccessToken parameter names
Fixes #40

I'm actually cleaning up the other function parameters as they should be different than the more global VB "constants".  I don't have an easy way to test that this works, so I'm going to create a PR.  If someone can point me to instructions on how to test in the future, I'll gladly give it a shot.