Fix java server-api sample to use correct Base64 instance
The init method first created a Base64 instance with urlsafe flag set to true. But then it used the *static* decodeBase64 method that actually creates a new Base64 instance with the urlsafe flag set to false. This fixes the code to use an instance method on the Base64 class instead.