First two MFA use cases
MFA for Employees and MFA for Customers. Each has 4 graphics which have
placed at /media/articles/email-wall/use-cases/mfa-for-employees/ and
/media/articles/email-wall/use-cases/mfa-for-customers/ respectively.
There is a link for a PDF at the top of each .md file, but unclear where
I should put the PDFs, so I'll keep them in the same dir as the .md file
it corresponds to.