Create index.md
Auth0 provides multiple locations for storing different types of data associated with authenticating an app’s users. The purpose of this document is to demonstrate the best practices in using these storage mechanisms efficiently and securely and to give you a look at an end-to-end experience through an application connected to both auth0 and an external database.