Update CONTRIBUTING.md
I have edited only a small fragment of this document, from "headings" to "URLs" section. 
There were several errors found, mostly phrases in the opposite order. It seems to me that the original writer is not an english native speaker nor skilled at the language. The text has been written as a literal translation from Spanish or another related language.
Also, there is a lack of punctuation symbols and the writer has had a lack of attention during the writing process.
Kindest regards,
Carolina