Enhance guidance for multiple DB connections
There currently is a limit with Lock today where it will only support the first 50 enabled database connections for a given client. This is because Lock loads connection metadata from the Auth0 CDN and to avoid those files from becoming ridiculously large, a limit was employed. Customers should be aware of this and consider the one client per tenant approach described in the WARNING included in this PR.