Update Amazon Cognito doc
This doc was out of date using Swift code samples that are no longer
valid with the service. Instead of updating the sample just for Swift I
linked to the AWS docs on how to implement this for each platform so
this doc does not need to keep being updated.

https://trello.com/c/8B41PWJ9/1143-update-amazon