OIDC Dyn Reg endpoint - Optional client_name
Update the documentation for the `client_name` attribute to reflect that it is optional. When omitted by the caller, the default value "My App" is used.

Also, add link to the Tenant Settings page where a user can enable the "OIDC Dynamic Client Registration" feature. This comes as an alternative to setting it via management api.