Fix Golang jwks multi-key logic error
The present logic is very broken.  It steps through indexes (just "0") of the *certificate chain* of the first key (Keys[0].X5c) in the response.  It then uses those indexes to look at *each key in the response* (red flag!) and compare the kid attribute of that key, and on a match, it uses the *value offset* of the *certificate chain* to populate the match.

Using the range of one slice to walk through an entirely different slice can, of course, cause out-of-bounds crashes (if there was 1 key with 2 certificates in the chain, not that that is likely). 

The other big issues is that this doesn't actually handle multiple keys, e.g.
```{"keys":[{"kid":"mykey1", "x5c":["cert1"]},{"kid":"mykey2", "x5c":["cert2"]}]}```
e.g. during key rotations if two keys are present.  My proposed code now steps through the *keys*, and uses the first cert in the chain for the PEM (if there ever was a second certificate in the chain, it would be ignored, which I believe should be fine.)