Update index.md
The original language in the doc suggests that the act of granting a scope is equivalent to granting the associated permission to the caller. That is not the case. A scope only establishes the *subset* of user privileges that can be exercised by an app on behalf of its user. It doesn't grant permissions that aren't there. In other words: if App A has been granted Files.Read, the app will be able to read files as long as the current user does have the corresponding privilege at the resource. If the user doesn't have file read privileges, the app will NOT be able to read files even if its access token contains the scope Mail.Read. 
For more details see https://twitter.com/vibronet/status/923214683124875264