API AuthZ landing page: add clarification on this article's purpose (#6299)
* add clarification on this article's purpose
We received the following negative feedback on this page: `not an article but a collection of links`
This is what this is meant to be, a landing page for API Authorization listing our relevant content. I added one sentence that clarifies that (and did some copyediting along the way).

* remove redundant links: 
We link to the "Introducing OIDC Conformant Authentication" article from the banner at the top of this page. We also link to the "OIDC Conformant Authentication Adoption Guide" from the "Introducing OIDC Conformant Authentication" article. Removing these two links since we already have a lot there and I see no need to duplicate the links.