remove client id from URL
If you don’t add the client id you’ll get the metadata from the global client (the cert and other info from the tenant). If you add it you get the ones from a particular client. We would like to have only the one from the tenant, so we are removing the guidance that says otherwise. Any requests to support the client id should be forwarded to crew-auth, alongside with info on the use case from the customer.