Reword considerations & add caveat for appliance
Main purpose for this PR is to add the note about appliance customers not being able to do a tenant restoration. This was brought to our attention on the Sustainment team by Nelly Delgado. We're updating our internal guides for this as well. 

This might need to pass through additional review before being approved. Tenant restoration is not a fast or easy task to complete and does come with some risk.