Fix bias in randomString()
Charset `[c % charset.length]` technically works for mapping a random value to a character. But by using the modulo operator, a possible bias in introduced. In this case, with the length of the charset being 65, the last 4 characters are chosen less often than the others.

Another possible solution for removing the bias, is using a charset of length 64. However, this is very error-prone because a user could easily change the charset.

Note: as long as the used nonce is long enough, using the biased implementation won't easily cause any vulnerabilities. However, the name of the function and the accompanying text suggest it can be used to generate a secure random strings, which is not true.