Update login for React quickstart to React Router 4
There was some incorrect code towards the end of this file describing how to use a callback function, which used Route's render props facility to call a '`handleAuthentication`' method which tests the hash and in turn call the `auth.handleAuthentication` method. This appeared to be an artefact from React Router 3's implementation of `onEnter`, though I'm not sure. In any case, the function would not have been called with the has argument as described, although the `auth.handleAuthentication` function would have been unaffected. I have rewritten it to pass the testing function - which I have renamed `testAuthHash` to distinguish it from the auth version - to the `Callback` component, where it is called on `componentDidMount`, which is the recommended way to handle this sort of thing in React Router 4.