Creatd user-export usage details
Initial commit, need few formatting to be done. No more technical data needs to go in this file. This information is not present in the public documentation and customers like Telkomsel use these steps. Everytime when they want to get user-exports they create a support ticket, So avoiding that by creating this document.