Fixed code snippet based on customer feedback
The error in the code snippet was noticed by a reader and reported
using the quickstart feedback tool. It is also out of alignment with the
GitHub sample, so this PR corrects that error.

Also disabled some markdownlint warnings regarding headers, so that the
page doesn't have to be rearchitected. It is due to be rewritten shortly
anyway.