Adding tip to best practices about not storing credentials in metadata
A number of customers have either inquired about or are actually storing access tokens and refresh tokens in app_metadata or user_metadata. This is not a good practice, as it enables any dashboard administrator to view and copy those tokens. So I am suggesting this update to explicitly state this is not recommended.