Removed SDK installation instructions from QS intro
These have proven to be confusing for the customer, as they are generic
and don't relate to the quickstart tutorial at all. One of the main
pieces of feedback is that they see the import statement but they don't
know which file to put it in.

Installation of the library is repeated further down the tutorial in
context, and more detailed instructions are included as part of the
README on GitHub.

Also removed the "Authenticating with Auth0" include as it's not hugely
required knowledge for the new SDK and just adds fluff.