Clarifying What the User Should See When Testing (#8629)
After completing step 24, the documentation says that the "expenses" page should load successfully. It does load successfully but the pages displays the text, "loading...". This is slightly confusing since an indefinite loading text hints at something is not working. I think it is worth it to specify to the user that the page will open successfully AND they can expect to see the loading text.