Update 02-calling-an-api.md to explain consent
There is currently no mention of why consent is required and following this example app only leads to confusion. It took WAY to long to find that the management API requires consent regardless of anything and/or why/when consent is explicitly necessary when using a SPA. These instructions seem to reflect a 3rd party app, rather than a first party app.