Update QuickStart Python Doc
Added the setting of the scope to the client.
While following the quickstart guide for python, when requesting an
access token the scope `read:messages` was not included in the response.

After investigating, the defined screen to assign permissions, was
actually a list of supported permissions for the API, but not the scopes
granted to the client when requesting an access token.

Added that extra step to the quickstart